+++ src/test/java/dk/kb/present/ViewTest.java
@@ -1,7 +1,8 @@
 package dk.kb.present;
 
 import dk.kb.present.config.ServiceConfig;
-import dk.kb.present.holdback.HoldbackDatePicker;
+import dk.kb.present.dr.holdback.HoldbackDatePicker;
+import dk.kb.present.dr.restrictions.ProductionIdLookup;
 import dk.kb.storage.model.v1.DsRecordDto;
 import dk.kb.util.Resolver;
 import dk.kb.util.yaml.YAML;
@@ -43,6 +44,7 @@ class ViewTest {
         try {
             ServiceConfig.initialize("conf/ds-present-behaviour.yaml", "internal-test-setup.yaml");
             config = ServiceConfig.getConfig();
+            ProductionIdLookup.init();
         } catch (IOException e) {
             fail();
         }
@@ -313,6 +315,22 @@ class ViewTest {
 
         String solrdoc = jsonldView.apply(recordDto);
         assertTrue(solrdoc.contains("\"dr_production_id\":\"8030782300\""));
+
+        assertTrue(solrdoc.contains("\"dr_id_restricted\":\"false\""));
+    }
+
+    @Test
+    @Tag("integration")
+    void testRestrictedDRProductionID() throws IOException {
+        HoldbackDatePicker.init();
+        View jsonldView = getSolrTvViewForPreservicaRecord();
+        String pvica = Resolver.resolveUTF8String(TestFiles.PVICA_DOMS_MIG_968dd87e);
+        DsRecordDto recordDto = new DsRecordDto().data(pvica).id("test.id").mTimeHuman("2023-11-29 13:45:49+0100").mTime(1701261949625000L)
+                .origin("ds.tv").kalturaId("randomKalturaId");
+
+        String solrdoc = jsonldView.apply(recordDto);
+
+        assertTrue(solrdoc.contains("\"dr_id_restricted\":\"true\""));
     }
 
     @Test